do nothing if latest version is already installed.

+++ update.sh
@@ -5,24 +5,28 @@ exit_code=$?
 
 if [ $exit_code == 0 ]; then
   read -r ver url <<<$output
-  wget $url -O minecraft_server.jar.$ver
+  if [ ! $(readlink minecraft_server.jar) == minecraft_server.jar.$ver ]; then
+    wget $url -O minecraft_server.jar.$ver
 
-  if [ $(readlink minecraft_server.jar.prev) ]; then
-    archive_ver=$(readlink minecraft_server.jar.prev)
-    unlink minecraft_server.jar.prev
-  if [ ! -d "archive" ]; then
-    mkdir archive
-  fi
-  mv $archive_ver archive/.
-  fi
+    if [ $(readlink minecraft_server.jar.prev) ]; then
+      archive_ver=$(readlink minecraft_server.jar.prev)
+      unlink minecraft_server.jar.prev
+      if [ ! -d "archive" ]; then
+        mkdir archive
+      fi
+      mv $archive_ver archive/.
+    fi
 
-  if [ $(readlink minecraft_server.jar) ]; then
-    replace_ver=$(readlink minecraft_server.jar)
-    unlink minecraft_server.jar
-    ln -s $replace_ver minecraft_server.jar.prev
-  fi
+    if [ $(readlink minecraft_server.jar) ]; then
+      replace_ver=$(readlink minecraft_server.jar)
+      unlink minecraft_server.jar
+      ln -s $replace_ver minecraft_server.jar.prev
+    fi
 
-  ln -s minecraft_server.jar.$ver minecraft_server.jar
+    ln -s minecraft_server.jar.$ver minecraft_server.jar
+  else
+    echo Server already on latest version
+  fi
 else
   echo $output
 fi
